Improved extensibiility
The classes should allow for more flexibility in the kinds of
calculations it can do. It still needs work to ensure that the total
chemical potential, mu, and all associated variables get set, and
variables may need to be moved between classes to make handling them
less awkward, but it should be relatively easy to implement arbitrary
chemical potentials and to sweep different parameters (not just
temperature).